refactor(vox): consolidate autocomplete handlers into single file
Move FvoxClipAutocompleteHandler and HgruntClipAutocompleteHandler into
VoxClipAutocompleteHandler.cs for better organization.
